+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.2] - 2022-03-21
+
+### Fixed
+
+- `/private` now passes the `hobbies`, `links`, and `privateInformation` fields that are required by child components
+- `/private` will now pass the secret key to the PDF link
+- PDF generator can now receive a `secret` query param to correctly list private fields
+
 ## [2.0.1] - 2022-02-26
 
 ### Added